Converted to JSON config file
